requirements.txt
aiohttp==2.2.5
async-timeout==1.4.0
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-GraphQL==1.4.1
graphene==2.0.dev20170802065539
graphql-core==2.0.dev20170801051721
graphql-relay==0.4.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
multidict==3.2.0
promise==2.1.dev0
six==1.11.0
typing==3.6.2
Werkzeug==0.12.2
yarl==0.12.0
